change to getting inserted id

+++ domain/repository/todoItemRepository.go
@@ -7,6 +7,6 @@ import (
 
 type TodoItemRepository interface {
 	GetTodoItems(DB *sql.DB) ([]*domain.TodoItem, error)
-	Insert(DB *sql.DB, name string) (int64, error)
+	Insert(DB *sql.DB, name string) (int, error)
 	Delete(DB *sql.DB, ids []string) error
 }
